4.1. Продолжение знакомства с SQL Доработки: исправил тип на double

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -72,7 +72,7 @@ public class StudentController {
     }
     @GetMapping("/average-age")
     @Operation(summary = "Get average age students")
-    public int getAvgAgeStudents() {
+    public double getAvgAgeStudents() {
         return studentService.getAvgAgeStudents();
     }
     @GetMapping("/get-last-five")
